+++ src/plugins/external_tool_support/src/tcoffee/TCoffeeSupportTask.cpp
@@ -77,8 +77,7 @@ void TCoffeeSupportTask::prepare(){
             emit si_stateChanged();
         }
     }
-    tmpDir.cd(AppContext::getAppSettings()->getUserAppsSettings()->getTemporaryDirPath());
-    if(!tmpDir.mkdir(tmpDirName)){
+    if(!tmpDir.mkpath(AppContext::getAppSettings()->getUserAppsSettings()->getTemporaryDirPath()+"/"+tmpDirName)){
         stateInfo.setError(tr("Can not create directory for temporary files."));
         emit si_stateChanged();
         return;
@@ -159,13 +158,15 @@ QList<Task*> TCoffeeSupportTask::onSubTaskFinished(Task* subTask) {
 }
 Task::ReportResult TCoffeeSupportTask::report(){
     //Remove subdir for temporary files, that created in prepare
-    QDir tmpDir(QFileInfo(url).absoluteDir());
-    foreach(QString file, tmpDir.entryList()){
-        tmpDir.remove(file);
-    }
-    if(!tmpDir.rmdir(tmpDir.absolutePath())){
-        stateInfo.setError(tr("Can not remove directory for temporary files."));
-        emit si_stateChanged();
+    if(!url.isEmpty()){
+        QDir tmpDir(QFileInfo(url).absoluteDir());
+        foreach(QString file, tmpDir.entryList()){
+            tmpDir.remove(file);
+        }
+        if(!tmpDir.rmdir(tmpDir.absolutePath())){
+            stateInfo.setError(tr("Can not remove directory for temporary files."));
+            emit si_stateChanged();
+        }
     }
 
     return ReportResult_Finished;
